# apps/task/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "summary": "",
        "description": "Create a new Task. The user must have Create permission for Tasks.\nUser can only create Tasks for its own Projects.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTaskDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      },
      "get": {
        "operationId": "TasksController_findAll",
        "summary": "",
        "description": "Get a list of all Tasks. The user must have permission to Read all Tasks.",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "The project to which the tasks belongs.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/activate": {
      "put": {
        "operationId": "TasksController_activate",
        "summary": "",
        "description": "By default, Task is created as draft. Use this endpoint to Activate the task.\nActivating a task changes its status to open, and it can no longer be edited.\nThe calling user must have Update permission on the Task.\nThere are certain pre-requisites for activating a Task.\n\n1. The Task must contain Questions\n2. The Task must contain crowd selection criteria (Skills and Countries)\n3. The Task must be in Draft state\n4. The task must contain Pictures, or Iframe Urls or textual description of Idea.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/close": {
      "put": {
        "operationId": "TasksController_close",
        "summary": "",
        "description": "Task can only be closed if it is in Open state.\nThe calling user must have Update permission on the Task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/iterations/{projectId}": {
      "get": {
        "operationId": "TasksController_findIterations",
        "summary": "",
        "description": "Get a list of all task iterations for a projectId. The iterations are returned in descending order.\nThe user must have Read permission on the Project and its Task iterations.",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "The project id to list iterations for",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/open": {
      "get": {
        "operationId": "TasksController_findOpenTasks",
        "summary": "",
        "description": "Get open tasks for Crowdworker to work on. The calling user must have Create permission for the Feedback,\nbecause only those users can see open tasks which have permission to create feedbacks.\nOnly those Tasks will be returned where user have not provided Feedback yet.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "summary": "",
        "description": "Get a Task. The calling user must have Read permission the Task either explicitly or implicitly.\nExplicit Permission: E.g Given by a role e.g developer role\nImplicit Permission: The task criteria matches calling user profile, and calling user have permission to create Feedback.\nImplicit Permission: This user has created Feedback for this Task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"]
      },
      "patch": {
        "operationId": "TasksController_update",
        "summary": "",
        "description": "Update a Task. The calling user must have Update permission the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTaskDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "summary": "",
        "description": "Delete a Task. The calling user must have Delete permission the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/feedback-stats": {
      "get": {
        "operationId": "TasksController_feedbackStats",
        "summary": "",
        "description": "Get statistics for feedbacks e.g, no of each star rating answer, number of thumbs up/down, no of each radio option selections.\nThe user must have Read Permission for the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackStatsResponseDto"
                }
              }
            }
          }
        },
        "tags": ["tasks"]
      }
    },
    "/tasks/{id}/image-upload-signature": {
      "get": {
        "operationId": "TasksController_imageUploadSignature",
        "summary": "",
        "description": "Images are uploaded to cloudinary.com. For uploading an image, a signature is required.\nThis endpoint returns that signature which clients can use to upload images.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadSignatureResponseDto"
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/batch-create-images": {
      "post": {
        "operationId": "TasksController_batchCreateImages",
        "summary": "",
        "description": "Batch create Images\nThe user must have Update permission for Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchCreateImagesDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/images": {
      "get": {
        "operationId": "TasksController_findAllImages",
        "summary": "",
        "description": "Get list of images for the Task.\nThe caller must have read permission for the Task either explicitly or either implicitly.\nExplicit Permission: E.g Given by a role e.g developer role\nImplicit Permission: The task's criteria matches calling user profile, and calling user have permission to create Feedback.\nImplicit Permission: This user has created Feedback for this Task",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "prototypeNumber",
            "required": false,
            "in": "query",
            "description": "Optional filtering by prototypeNumber",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Image" }
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/images/{imageId}": {
      "delete": {
        "operationId": "TasksController_removeImage",
        "summary": "",
        "description": "Delete an Image for Task. The calling user must have Update permission on the Task.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/task-requests": {
      "get": {
        "operationId": "TasksController_findTaskRequests",
        "summary": "",
        "description": "Get a list of all TaskRequests for a taskId.\nThe task must belong to the user.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TaskRequest" }
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
    "/tasks/{id}/file-upload-signature": {
      "get": {
        "operationId": "TasksController_fileUploadSignature",
        "summary": "",
        "description": "Files are uploaded to cloudinary.com. For uploading a file, a signature is required.\nThis endpoint returns that signature which clients can use to upload files.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadSignatureResponseDto"
                }
              }
            }
          }
        },
        "tags": ["tasks"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/feedbacks/tasks/{taskId}": {
      "get": {
        "operationId": "FeedbacksController_findFeedbacks",
        "summary": "",
        "description": "Get all Feedbacks for a Task. The user must have Read permission on the Task.",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "description": "The task id to list Feedbacks for",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Feedback" }
                }
              }
            }
          }
        },
        "tags": ["feedbacks"],
        "security": [{ "bearer": [] }]
      }
    },
...
  },
  "info": {
    "title": "Task",
    "description": "Task service is responsible for iterative validation of Prototypes. Tasks are created and Feedback is provided through this service.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "Bearer",
        "description": "Authentication is done by a signed JWT",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "FileUploadSignatureResponseDto": {
        "type": "object",
        "properties": {
          "apiKey": { "type": "string" },
          "uploadUrl": { "type": "string" },
          "signature": { "type": "string" },
          "timestamp": { "type": "number" },
          "folder": { "type": "string" }
        },
        "required": ["apiKey", "uploadUrl", "signature", "timestamp", "folder"]
      },
...
      "Question": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string" },
          "type": {
            "type": "string",
            "enum": ["text", "radio", "star-rating", "thumbs-rating"]
          },
          "order": { "type": "number", "minimum": 1 },
          "radioOptions": { "type": "array", "items": { "type": "string" } },
          "taskId": { "type": "string" }
        },
        "required": [
          "id",
          "description",
          "type",
          "order",
          "radioOptions",
          "taskId"
        ]
      },
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "testType": { "type": "string", "enum": ["basic", "comparison"] },
          "prototypeFormat": {
            "type": "string",
            "enum": ["image", "iframe", "text"]
          },
          "accessType": {
            "type": "string",
            "enum": ["open", "nda", "request"]
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number", "minimum": 0 },
          "incentive": { "type": "number", "minimum": 0 },
          "projectId": { "type": "string" },
          "skills": { "type": "array", "items": { "type": "string" } },
          "countries": { "type": "array", "items": { "type": "string" } },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          }
        },
        "required": [
          "title",
          "testType",
          "prototypeFormat",
          "accessType",
          "budget",
          "incentive",
          "projectId",
          "skills",
          "countries",
          "questions"
        ]
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "url": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "prototypeNumber": { "type": "number", "minimum": 1, "maximum": 2 },
          "taskId": { "type": "string" },
          "cloudId": { "type": "string" }
        },
        "required": [
          "id",
          "url",
          "dateCreated",
          "prototypeNumber",
          "taskId",
          "cloudId"
        ]
      },
      "Skill": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["id", "name"]
      },
      "Country": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "code": { "type": "string" }
        },
        "required": ["id", "name", "code"]
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "testType": { "enum": ["basic", "comparison"], "type": "string" },
          "prototypeFormat": {
            "enum": ["image", "iframe", "text"],
            "type": "string"
          },
          "status": { "enum": ["draft", "open", "closed"], "type": "string" },
          "accessType": {
            "enum": ["open", "nda", "request"],
            "type": "string"
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number" },
          "incentive": { "type": "number" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "images": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Image" }
          },
          "skills": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Skill" }
          },
          "countries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Country" }
          },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          },
          "feedbacks": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Feedback" }
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "dateCreated",
          "testType",
          "prototypeFormat",
          "status",
          "accessType",
          "iframeUrl1",
          "iframeUrl2",
          "textualDescription1",
          "textualDescription2",
          "minAge",
          "maxAge",
          "minExperience",
          "maxExperience",
          "budget",
          "incentive",
          "userId",
          "projectId",
          "images",
          "skills",
          "countries",
          "questions",
          "feedbacks"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "textAnswer": { "type": "string" },
          "radioAnswer": { "type": "string" },
          "starRatingAnswer": { "type": "number", "minimum": 1, "maximum": 5 },
          "thumbsRatingAnswer": { "type": "string", "enum": ["up", "down"] },
          "feedbackId": { "type": "string" },
          "questionId": { "type": "string" }
        },
        "required": [
          "id",
          "textAnswer",
          "radioAnswer",
          "starRatingAnswer",
          "thumbsRatingAnswer",
          "feedbackId",
          "questionId"
        ]
      },
      "Feedback": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "comment": { "type": "string" },
          "paymentStatus": {
            "enum": ["pending", "completed"],
            "type": "string"
          },
          "feedbackRating": { "type": "number" },
          "feedbackRatingComment": { "type": "string" },
          "taskRating": { "type": "number" },
          "taskRatingComment": { "type": "string" },
          "taskId": { "type": "string" },
          "task": { "$ref": "#/components/schemas/Task" },
          "userId": { "type": "string" },
          "answers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Answer" }
          }
        },
        "required": [
          "id",
          "dateCreated",
          "comment",
          "paymentStatus",
          "feedbackRating",
          "feedbackRatingComment",
          "taskRating",
          "taskRatingComment",
          "taskId",
          "task",
          "userId",
          "answers"
        ]
      },
      "QuestionAnswerStats": {
        "type": "object",
        "properties": {
          "questionId": { "type": "string" },
          "questionType": {
            "type": "string",
            "enum": ["text", "radio", "star-rating", "thumbs-rating"]
          },
          "starRatingAnswer": { "type": "number" },
          "starRatingAnswerCount": { "type": "number" },
          "radioAnswer": { "type": "string" },
          "radioAnswerCount": { "type": "number" },
          "thumbsUpCount": { "type": "number" },
          "thumbsDownCount": { "type": "number" }
        },
        "required": [
          "questionId",
          "questionType",
          "starRatingAnswer",
          "starRatingAnswerCount",
          "radioAnswer",
          "radioAnswerCount",
          "thumbsUpCount",
          "thumbsDownCount"
        ]
      },
      "FeedbackStatsResponseDto": {
        "type": "object",
        "properties": {
          "stats": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/QuestionAnswerStats" }
          }
        },
        "required": ["stats"]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "testType": { "type": "string", "enum": ["basic", "comparison"] },
          "prototypeFormat": {
            "type": "string",
            "enum": ["image", "iframe", "text"]
          },
          "accessType": {
            "type": "string",
            "enum": ["open", "nda", "request"]
          },
          "iframeUrl1": { "type": "string" },
          "iframeUrl2": { "type": "string" },
          "textualDescription1": { "type": "string" },
          "textualDescription2": { "type": "string" },
          "minAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "maxAge": { "type": "number", "minimum": 18, "maximum": 67 },
          "minExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "maxExperience": { "type": "number", "minimum": 0, "maximum": 49 },
          "budget": { "type": "number", "minimum": 0 },
          "incentive": { "type": "number", "minimum": 0 },
          "skills": { "type": "array", "items": { "type": "string" } },
          "countries": { "type": "array", "items": { "type": "string" } },
          "questions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Question" }
          }
        }
      },
      "BatchCreateImagesDto": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Image" }
          }
        },
        "required": ["images"]
      },
      "TaskRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "dateCreated": { "format": "date-time", "type": "string" },
          "requestComment": { "type": "string" },
          "rejectionComment": { "type": "string" },
          "ndaUrl": { "type": "string" },
          "ndaCloudId": { "type": "string" },
          "requestStatus": {
            "enum": ["accepted", "requested", "rejected"],
            "type": "string"
          },
          "userId": { "type": "string" },
          "taskId": { "type": "string" }
        },
        "required": [
          "id",
          "dateCreated",
          "requestComment",
          "rejectionComment",
          "ndaUrl",
          "ndaCloudId",
          "requestStatus",
          "userId",
          "taskId"
        ]
      },
...
    }
  }
}